+++ packages/nesd/lib/nes/cartridge/mapper/mapper.dart
@@ -9,6 +9,7 @@ import 'package:nesd/nes/cartridge/mapper/bandai_fcg.dart';
 import 'package:nesd/nes/cartridge/mapper/br909x.dart';
 import 'package:nesd/nes/cartridge/mapper/cnrom.dart';
 import 'package:nesd/nes/cartridge/mapper/dma_settings.dart';
+import 'package:nesd/nes/cartridge/mapper/fme7.dart';
 import 'package:nesd/nes/cartridge/mapper/gxrom.dart';
 import 'package:nesd/nes/cartridge/mapper/mapper176.dart';
 import 'package:nesd/nes/cartridge/mapper/mapper45.dart';
@@ -82,6 +83,7 @@ abstract class Mapper {
       30 => UNROM512(subMapperId),
       45 => Mapper45(subMapperId),
       66 => GxROM(),
+      69 => FME7(),
       71 => BR909x(),
       118 => TxSROM(),
       176 => Mapper176(subMapperId),