+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Added
+- Sunsoft FME-7 (mapper 69) support
 - Auto save now also saves when quitting a game, quitting NESd, or sending it to the background
 - The Previous Tab and Next Tab actions now page through the recent ROM and save state grids